refactor: co-locate use*Deposit hooks
